requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
alembic==1.16.4
annotated-types==0.7.0
anyio==4.9.0
apache-airflow==2.10.0
apache-airflow-providers-common-compat==1.7.2
apache-airflow-providers-common-io==1.6.1
apache-airflow-providers-common-sql==1.27.3
apache-airflow-providers-fab==1.5.3
apache-airflow-providers-ftp==3.13.1
apache-airflow-providers-http==5.3.2
apache-airflow-providers-imap==3.9.1
apache-airflow-providers-smtp==2.1.1
apache-airflow-providers-sqlite==4.1.1
apispec==6.8.2
appdirs==1.4.4
argcomplete==3.6.2
asgiref==3.9.1
async-timeout==5.0.1
attrs==25.3.0
babel==2.17.0
blinker==1.9.0
cachelib==0.13.0
cachetools==5.5.2
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
clickclick==20.10.2
cloudpickle==3.1.1
colorama==0.4.6
colorlog==6.9.0
ConfigUpdater==3.2
connexion==2.14.2
contourpy==1.3.2
cron-descriptor==1.4.5
croniter==6.0.0
cryptography==45.0.5
cycler==0.12.1
dacite==1.9.2
databricks-sdk==0.58.0
Deprecated==1.2.18
deprecation==2.1.0
dill==0.4.0
distro==1.9.0
dnspython==2.7.0
docker==7.1.0
dynaconf==3.2.11
email_validator==2.2.0
evidently==0.7.9
exceptiongroup==1.3.0
Faker==37.4.2
fastapi==0.116.1
filelock==3.18.0
Flask==2.2.5
Flask-AppBuilder==4.5.3
Flask-Babel==2.0.0
Flask-Caching==2.3.1
Flask-JWT-Extended==4.7.1
Flask-Limiter==3.12
Flask-Login==0.6.3
Flask-Session==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.2.2
fonttools==4.58.5
frozenlist==1.7.0
fsspec==2025.7.0
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.40.3
google-re2==1.1.20240702
googleapis-common-protos==1.70.0
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
greenlet==3.2.3
grpcio==1.73.1
gunicorn==23.0.0
h11==0.16.0
htmlmin==0.1.12
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
ImageHash==4.3.1
importlib_metadata==8.7.0
inflection==0.5.1
iterative-telemetry==0.0.10
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
lazy-object-proxy==1.11.0
lightgbm==4.6.0
limits==5.4.0
linkify-it-py==2.0.3
litestar==2.16.0
litestar-htmx==0.5.0
llvmlite==0.44.0
lockfile==0.12.2
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
marshmallow-oneofschema==3.2.0
marshmallow-sqlalchemy==0.28.2
matplotlib==3.10.0
mdit-py-plugins==0.4.2
mdurl==0.1.2
methodtools==0.4.7
mlflow==3.1.1
mlflow-skinny==3.1.1
more-itertools==10.7.0
msgspec==0.19.0
multidict==6.6.3
multimethod==1.12
multipart==1.2.1
mypy_extensions==1.1.0
networkx==3.4.2
nltk==3.9.1
numba==0.61.0
numpy==1.25.2
nvidia-nccl-cu12==2.27.6
opentelemetry-api==1.35.0
opentelemetry-exporter-otlp==1.35.0
opentelemetry-exporter-otlp-proto-common==1.35.0
opentelemetry-exporter-otlp-proto-grpc==1.35.0
opentelemetry-exporter-otlp-proto-http==1.35.0
opentelemetry-proto==1.35.0
opentelemetry-sdk==1.35.0
opentelemetry-semantic-conventions==0.56b0
ordered-set==4.1.0
packaging==25.0
pandas==2.0.3
pathspec==0.12.1
patsy==1.0.1
pendulum==3.1.0
phik==0.12.4
pillow==11.3.0
plotly==5.24.1
pluggy==1.6.0
polyfactory==2.22.1
prison==0.2.1
propcache==0.3.2
protobuf==6.31.1
psutil==7.0.0
psycopg2-binary==2.9.10
puremagic==1.30
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
pyparsing==3.2.3
python-daemon==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-nvd3==0.16.0
python-slugify==8.0.4
pytz==2025.2
PyWavelets==1.8.0
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rich==13.9.4
rich-argparse==1.7.1
rich-click==1.8.9
rpds-py==0.26.0
rsa==4.9.1
scikit-learn==1.7.0
scipy==1.15.3
seaborn==0.13.2
setproctitle==1.3.6
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==1.4.54
SQLAlchemy-JSONField==1.0.2
SQLAlchemy-Utils==0.41.2
sqlparse==0.5.3
starlette==0.47.1
statsmodels==0.14.5
tabulate==0.9.0
tenacity==9.1.2
termcolor==3.1.0
text-unidecode==1.3
threadpoolctl==3.6.0
tomli==2.2.1
tqdm==4.67.1
typeguard==4.4.4
typer==0.16.0
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
uc-micro-py==1.0.3
ujson==5.10.0
unicodecsv==0.14.1
universal_pathlib==0.2.6
urllib3==2.5.0
uuid6==2025.0.1
uvicorn==0.35.0
uvloop==0.21.0
visions==0.8.1
watchdog==6.0.0
watchfiles==1.1.0
websockets==15.0.1
Werkzeug==2.2.3
wirerope==1.0.0
wordcloud==1.9.4
wrapt==1.17.2
WTForms==3.2.1
xgboost==3.0.2
yarl==1.20.1
ydata-profiling==4.16.1
zipp==3.23.0
